Agents/seo-linkedin-optimization-agent/requirements.txt
# SEO LinkedIn Optimization Agent Requirements
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
pydantic==1.8.2